refactor: modularize abilities into individual files
Split monolithic Abilities.php (1318 lines) into clean modular structure:

inc/Core/Abilities/
  Registry.php          — Registers category + all abilities
  AuthStatus.php        — auth-status ability
  SaveCredentials.php   — save-credentials ability
  Disconnect.php        — disconnect ability
  PostFromMedia.php     — post-from-media + image processing helpers
  ListMedia.php         — list-media ability
  PostNow.php           — post-now ability
  SchedulePost.php      — schedule-post ability
  GetScheduledPosts.php — get-scheduled-posts ability
  PostStatus.php        — post-status ability

Each ability class has get_definition() and execute() methods.
Registry orchestrates registration via WordPress Abilities API hooks.
REST endpoints updated to call modular classes directly.
PSR-4 autoloader handles new namespace paths automatically.
