+++ post-to-instagram.php
@@ -97,7 +97,7 @@ function pti_init_plugin() {
     PostToInstagram\Core\Actions\Post::register();
     PostToInstagram\Core\Actions\Schedule::register();
     PostToInstagram\Core\Actions\Cleanup::register();
-    PostToInstagram\Core\Abilities::register();
+    PostToInstagram\Core\Abilities\Registry::register();
 
     if ( is_admin() ) {
         PostToInstagram\Core\Admin::register();
